Refactor: Implements `Mapper` and `DTO`

+++ src/main/java/com/carloslonghi/ScrumHub/service/EmployeeService.java
@@ -1,5 +1,7 @@
 package com.carloslonghi.ScrumHub.service;
 
+import com.carloslonghi.ScrumHub.dto.EmployeeDTO;
+import com.carloslonghi.ScrumHub.mapper.EmployeeMapper;
 import com.carloslonghi.ScrumHub.model.EmployeeModel;
 import com.carloslonghi.ScrumHub.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class EmployeeService {
@@ -14,28 +17,42 @@ public class EmployeeService {
     @Autowired
     private EmployeeRepository employeeRepository;
 
-    public List<EmployeeModel> getAllEmployees() {
-        return employeeRepository.findAll();
+    @Autowired
+    private EmployeeMapper employeeMapper;
+
+    public List<EmployeeDTO> getAllEmployees() {
+        List<EmployeeModel> employees = employeeRepository.findAll();
+        return employees.stream()
+                .map(employeeMapper::map)
+                .collect(Collectors.toList());
     }
 
-    public EmployeeModel createEmployee(EmployeeModel employee) {
-        return employeeRepository.save(employee);
+    public EmployeeDTO createEmployee(EmployeeDTO employeeDTO) {
+        EmployeeModel employee = employeeMapper.map(employeeDTO);
+        employee = employeeRepository.save(employee);
+        return employeeMapper.map(employee);
     }
 
-    public EmployeeModel getEmployeeById(Long id) {
+    public EmployeeDTO getEmployeeById(Long id) {
         Optional<EmployeeModel> employee = employeeRepository.findById(id);
-        return employee.orElse(null);
+        return employee.map(employeeMapper::map).orElse(null);
     }
 
-    public EmployeeModel updateById(Long id, EmployeeModel employee) {
-        if (employeeRepository.existsById(id)) {
-            employee.setId(id);
-            return employeeRepository.save(employee);
+    public EmployeeDTO updateById(Long id, EmployeeDTO employeeDTO) {
+        Optional<EmployeeModel> employeeExist = employeeRepository.findById(id);
+        if (employeeExist.isPresent()) {
+            EmployeeModel employeeUpdated = employeeMapper.map(employeeDTO);
+            employeeUpdated.setId(id);
+            EmployeeModel employee = employeeRepository.save(employeeUpdated);
+            return employeeMapper.map(employee);
         }
         return null;
     }
 
     public void deleteEmployeeById(Long id) {
-        employeeRepository.deleteById(id);
+        Optional<EmployeeModel> employeeExist = employeeRepository.findById(id);
+        if (employeeExist.isPresent()) {
+            employeeRepository.deleteById(id);
+        }
     }
 }
